// src/pages/register/student/payment/type.ts
export type StudentPaymentType = {
  january: {
    value: number;
    interest: number;
  };
  february: {
    value: number;
    interest: number;
  };
  march: {
    value: number;
    interest: number;
  };
  firstQuarter: {
    value: number;
    interest: number;
  };
  april: {
    value: number;
    interest: number;
  };
  may: {
    value: number;
    interest: number;
  };
  june: {
    value: number;
    interest: number;
  };
  secondQuarter: {
    value: number;
    interest: number;
  };
  july: {
    value: number;
    interest: number;
  };
  august: {
    value: number;
    interest: number;
  };
  september: {
    value: number;
    interest: number;
  };
  thirtQuarter: {
    value: number;
    interest: number;
  };
  october: {
    value: number;
    interest: number;
  };
  november: {
    value: number;
    interest: number;
  };
  december: {
    value: number;
    interest: number;
  };
  fourthQuarter: {
    value: number;
    interest: number;
  };
  total: {
    value: number;
    interest: number;
  };
};